+++ src/pages/Secure.tsx
@@ -1,7 +1,7 @@
 import firebase from 'firebase/app'
 // import 'firebase/storage'
 // import 'firebase/functions'
-import { useState, useEffect } from 'react'
+import { useState, useEffect, useRef } from 'react'
 
 // const storage = firebase.storage()
 
@@ -14,22 +14,33 @@ import { useState, useEffect } from 'react'
 
 const Secure = () => {
   const [token, setToken] = useState('')
-  // const [error, setError] = useState('')
+  const imgEl = useRef(null)
+  const [authorized, setAuthorized] = useState(false)
 
   useEffect(() => {
     firebase.auth().currentUser!.getIdToken(true).then((idToken) => {
-      console.log('>> idToken:', idToken)
+      // console.log('>> idToken:', idToken)
       setToken(idToken)
     }).catch((error) => {
       console.log('>> error:', error)
     })
   }, [])
 
+  function handleLoad () {
+    // const img = imgEl.current as unknown as HTMLImageElement
+    // console.log('>> imgLoaded:', img.naturalHeight)
+    setAuthorized(true) // handleLoad() only fires if user is authorized to see the asset
+  }
+
   return (
     <>
       An image will load if you're authorized to see it:<br /><br />
-      <img src={'http://localhost:5001/r002-cloud/us-central1/showPic_v0?token=' + token} />
-      {/* {error} */}
+      {!authorized &&
+        <>Sorry, you're not authorized to view this protected asset.</>
+      }
+      <img ref={imgEl} src={'https://us-central1-r002-cloud.cloudfunctions.net/showPic_v0?token=' + token}
+        style={authorized ? {} : { display: 'none' }}
+        onLoad={handleLoad} />
     </>
   )
 }